Aesthetic Update
Updates:
Added new icon for launcher. Credit to dexxaboy for the cool asset.
Added a copyright notice for the images (just in case).

Removed all executable data in favor of creating individual releases. It is much easier for those wanting to clone the repository to add their own modifications and is not as size-intensive; also, it is probably less confusing for those only wanting to use the application and not wanting to mess with the source files.